[IMP] Modularity: allow to specify properties to use when creating a cmis:object
